MailBot-ai-email-processor/requirements.txt
imaplib2
doctran
asyncio
pyyaml
python-dotenv
openai
langchain
langchain-community
langchain_google_genai